Normalize line endings
This git repo uses LF line endings in committed trees. But the prior .gitattributes setting told git to never regard any file as text so line endings would not be normalized. The result: depending on the configuration and/or OS of the git clone, line endings could be changed and an entire text file could appear changed in a PR, causing merge conflicts.

With this change, git will detect text files and normalize their line endings to LF to avoid this problem.